requirements.txt
ninja
pyyaml
pynvml
yacs
numpy
torch >= 1.13